beginning refactoring using package folders
